Shuffle Facebook classes
